Added a number of ways to create nptgroup object.
Added ability to specify a file containing a list of cell directories to use to create a nptgroup object.
Fixed ability to specify a search string to look for subdirectories in the current directory to use to create a nptgroup object.
Standarized on the use of Cell instead of using both Cell and Cluster to avoid confusion.