Made some minor changes to improve useability.
Now no longer checks markers at the top level, since it is unlikely the user will run ProcessLevel from a directory that should be skipped. Instead, the check was added to the subdirectories. It also now prints a message saying that the subdirectory was skipped.

Fixed bug that was causing LevelObject to not work properly.

Fixed bug that caused nptLevelCmd to not change directories properly.