Fixed bug in namePattern for channel names
Fixed bug in default namePattern for channel names. This bug was not noticed before as the information is usually obtained from the DPV configuration file. However, compiled programs in Matlab are not able to access the configuration file, so it has to fall back on the defaults specified in this function.