Changed defaults for names of data hierarchy
Changed defaults for names of data hierarchy to match what is being used in the lab. This was not necessary before since the information was obtained from the DPV configuration file. However, applications compiled with the Matlab Compiler are not able to access the config file, so the defaults in this function had to be changed.